pyproject.toml
[tool.poetry]
name = "e-market"
version = "0.1.0"
description = ""
authors = ["pestich <pestich@mail.ru>"]
readme = "README.md"
packages = [{include = "e_market"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.2"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
